% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/00_mcu.r
\docType{package}
\name{mcu-package}
\alias{mcu}
\alias{mcu-package}
\title{Model calibration utilities}
\description{
Utilities to aid in the calibration of models, especially those
with many parameters and significant computation times.
}

