% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensPlot.r
\name{sensPlot}
\alias{sensPlot}
\title{Plot results of sensitivity analysis}
\usage{
sensPlot(sout, by = c("fun", "par"), nc = 1)
}
\arguments{
\item{sout}{A list of data frames returned from a call to \code{\link{sens}}.}

\item{by}{Character string, either 'fun' or 'par'. If 'fun', the method
creates one plot for each component of the objective function. If
\code{sens} was called with a scalar-valued objective function (and thus
\code{sout} is of length 1) only a single plot is created.
If \code{by} is set to 'par', one plot is created for every parameter whose
influence was tested. This makes sense for vector-valued objective
functions only, i.e. if \code{sout} is of length > 1.}

\item{nc}{The number of columns used to arrange multiple plots. This info is
  passed to the \code{layout} function.}
}
\value{
\code{NULL}.
}
\description{
Plots the results of a sensitivity analysis carried out using \code{sens}.
}
\note{
If \code{by='fun'}, the plots can be interpreted as follows:
  \itemize{
    \item{} The value of the objective function called with default values
      for all parameters is plotted as a dot.
    \item{} The value of the objective function called with a particular
      parameter at the \emph{lower} limit is plotted as a \emph{blue} whisker line having
      its origin in the dot.
    \item{} The value of the objective function called with a particular
      parameter at the \emph{upper} limit is plotted as a \emph{red} whisker line having
      its origin in the dot.
  }

  If \code{by='par'}, the plots can be interpreted as follows:
  \itemize{
    \item{} The \emph{normalized} value of the objective function called
      with default values for all parameters is 1.
    \item{} The \emph{normalized} value of the objective function called
      with a particular parameter at the \emph{lower} limit is plotted as a \emph{blue}
      whisker line (having its origin at x=1).
    \item{} The \emph{normalized} value of the objective function called
      with a particular parameter at the \emph{upper} limit is plotted as a \emph{red}
      whisker line (having its origin at x=1).
  }
}
\examples{
# Sensitivity of parameters of a linear model
obs= data.frame(x=c(1,2), y=c(1,2))
model= function(p, x) { p["slope"] * x + p["intercept"] }
objfun= function(p, obs) { c(sse= sum((obs$y - model(p, obs$x))^2),
  mae= sum(abs(obs$y - model(p, obs$x)))) }
p= data.frame(
  name=c("slope","intercept"),
  default= c(1, 0.1),
  min= c(0.5, -1),
  max= c(2, 1)
)
s= sens(fn=objfun, p=p, obs=obs)
sensPlot(s, by="fun")
sensPlot(s, by="par")
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

