% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensPlotMatrix.r
\name{sensPlotMatrix}
\alias{sensPlotMatrix}
\title{Plot sensitivity info as a matrix}
\usage{
sensPlotMatrix(sout, thr = 0.9, xpars = TRUE)
}
\arguments{
\item{sout}{A list of data frames returned from a call to \code{\link{sens}}.}

\item{thr}{A number in range \eqn{0 < thr < 1}. This is used to pragmatically
distinguish between highly sensitive and less sensitive parameters.
\code{thr} specifies the minimum relative decrease in the value of the
objective function to consider a particular varied parameter as
\emph{highly} sensitive. A value of, say, 0.9 means that a parameter is
consired as highly sensitive if \eqn{ftst/fdef <= 0.9}, where \eqn{ftst}
and \eqn{fdef} denote the output of the objective function for a test value
and the default value of a parameter, respectively. Reasonable values are
probably between 0.8 and 0.95.}

\item{xpars}{Logical. Controls the plot's layout. If \code{TRUE}, the
  parameter names appear as column headers and the objective function(s) as
  row headers(s). If \code{FALSE}, the result matrix is transposed.}
}
\value{
\code{NULL}.
}
\description{
Plots the results of a sensitivity analysis carried out using \code{sens}
as a matrix of symbols. The output can facilitate manual model calibration
(see details below).
}
\note{
Symbols in the created graphics have the following meaning:
  \itemize{
    \item{} Triangles pointing up: Parameter should be increased to/beyond
      the tested upper limit in order to reduce the value of the objective
      function. Filled/non-filled: Relative sensitivity is high/low.
    \item{} Triangles pointing down: Parameter should be decreased to/beyond
      the tested lower limit in order to reduce the value of the objective
      function. Filled/non-filled: Relative sensitivity is high/low.
    \item{} Diamond: Lowest value of the objective function does not
      (exclusively) occur at a boundary of the tested parameter range. Hence,
      the optimum parameter values may be inside that range. Filled/non-filled:
      Relative sensitivity is high/low.
    \item{} Circle: Objective function is not sensitive to the parameter.
    \item{} Cross: No information due to non-finite return values of the
      objective function.
  }

  Note that the analysis does not account for possible parameter interactions
  such as compensation effects. For the example (see below), the created plot
  suggests that the intercept should be decreased (although the true optimum
  value is 0). This is due to the fact that all test values for the slope
  are actually too high (true optimum at 1).

  In cases with long names of parameters/functions, it will be necessary to
  adjust the plot margings accordingly (see example).
}
\examples{
# Sensitivity of parameters of a linear model
obs= data.frame(x=c(1,2), y=c(1,2))
model= function(p, x) { p["slope"] * x + p["intercept"] }
objfun= function(p, obs) { c(sse= sum((obs$y - model(p, obs$x))^2),
  mae= sum(abs(obs$y - model(p, obs$x)))) }
p= data.frame(
  name=c("slope","intercept"),
  default= c(1.5, 0.1),
  min= c(1.1, -1),
  max= c(2, 1)
)
s= sens(fn=objfun, p=p, obs=obs)
omar= par("mar")
par(mar=c(0.5,6,10,0.5))
sensPlotMatrix(sout=s, thr=0.75, xpars=TRUE)
par(mar=omar)
}
\author{
David Kneis \email{david.kneis@tu-dresden.de}
}

